app/requirements.txt
streamlit
pandas
plotly
faiss-cpu
databricks-sql-connector
sqlalchemy
pyarrow